+++ src/dev/Aziz/tilegame/worlds/World.java
@@ -6,23 +6,25 @@ import dev.Aziz.tilegame.entities.creatures.Enemy;
 import dev.Aziz.tilegame.entities.creatures.Player;
 import dev.Aziz.tilegame.entities.statics.TestEntity;
 import dev.Aziz.tilegame.items.ItemManager;
+import dev.Aziz.tilegame.tiles.Tile;
+import dev.Aziz.tilegame.tiles.TileManager;
 import dev.Aziz.tilegame.utils.Utils;
 
 import java.awt.*;
 
 public class World extends Layer{
 
-    public static final int WORLD_WIDTH = 50;
-    public static final int WORLD_HEIGHT = 50;
+
     public static final int SPAWN_X = 100;
     public static final int SPAWN_Y = 100;
 
-    private int width = WORLD_WIDTH, height = WORLD_HEIGHT;
-    private int spawnX = SPAWN_X, spawnY = SPAWN_Y;     //From the world.txt file
+    protected int spawnX = SPAWN_X, spawnY = SPAWN_Y;     //From the world.txt file
+
 
     //Entities
     private EntityManager entityManager;
     private ItemManager itemManager;
+    private TileManager tileManager;
 
 
     public EntityManager getEntityManager() {
@@ -31,6 +33,7 @@ public class World extends Layer{
 
     public World(Handler handler, String path){
         super(handler);
+        tileManager = new TileManager(handler);
 
         entityManager = new EntityManager(handler, new Player(handler, 0, 0));
         itemManager = new ItemManager(handler);
@@ -38,7 +41,7 @@ public class World extends Layer{
         entityManager.addEntity(new Enemy(handler,500, 100));
         entityManager.addEntity(new TestEntity(handler, 200, 200));
 
-        loadWorld(path, 2);
+        loadWorld(path, 0);
 
         entityManager.getPlayer().setX(spawnX);
         entityManager.getPlayer().setY(spawnY);
@@ -55,33 +58,37 @@ public class World extends Layer{
 
     public void render(Graphics g){
 
-        super.render(g);
+        tileManager.render(g);
         itemManager.render(g);
         entityManager.render(g);
 
     }
 
-
     private void loadWorld(String path, int index){
 
-        String file = Utils.loadXMLFileAsString(path, index);
-        String[] tokens = file.split(",");       //file.split(",") for comma     //file.split("\\s+")
-
-        tilesID = new int[width][height];
-
-        for(int y = 0; y < height; y++){
-            for(int x = 0; x < width; x++){
-                tilesID[x][y] = Utils.parseInt(tokens[(x + y * width)]);  //finding the corresponding data, basic math
-                System.out.print(tilesID[x][y] + " ");
-            }
-            System.out.println("\n");
-        }
+        loadLayer(path, index);
 
     }
 
-    public int getTilesID(int x, int y, int offset) {
-        return tilesID[x][y] + offset;
-    }
+
+    //private void loadWorld(String path, int index){
+
+    //    String file = Utils.loadXMLFileAsString(path, index);
+    //    String[] tokens = file.split(",");       //file.split(",") for comma     //file.split("\\s+")
+
+    //    tilesID = new int[width][height];
+
+    //    for(int y = 0; y < height; y++){
+    //        for(int x = 0; x < width; x++){
+    //            tilesID[x][y] = Utils.parseInt(tokens[(x + y * width)]);  //finding the corresponding data, basic math
+    //            System.out.print(tilesID[x][y] + " ");
+    //        }
+    //        System.out.println("\n");
+    //    }
+    //    System.out.println("Loading world finished!");
+
+    //}
+
 
     public int getWidth(){
         return width;
